convert shield to floats

+++ src/Components/ShieldComponent.hpp
@@ -4,18 +4,18 @@
 namespace Components {
 
 struct ShieldComponent {
-    unsigned int currentShield = 0;
-    unsigned int maxShield = 10;
+    float currentShield = 0.f;
+    float maxShield = 10.f;
     float regenRate = 1.f;
     float regenTimer = 0.f;
 
     ShieldComponent(){}
-    ShieldComponent(unsigned int currentShield, unsigned int maxShield, float regenRate) : currentShield(currentShield), maxShield(maxShield), regenRate(regenRate) {}
+    ShieldComponent(float currentShield, float maxShield, float regenRate) : currentShield(currentShield), maxShield(maxShield), regenRate(regenRate) {}
 
-    void setShield(unsigned int shield) { currentShield = shield; }
-    void incrementShield(unsigned int amount = 1) { currentShield += amount; }
-    void decrementShield(unsigned int amount = 1) { currentShield -= amount; }
-    unsigned int getShield() const { return currentShield; }
+    void setShield(float shield) { currentShield = shield; }
+    void incrementShield(float amount = 1.f) { currentShield += amount; }
+    void decrementShield(float amount = 1.f) { currentShield -= amount; }
+    float getShield() const { return currentShield; }
     };
 
     // TODO: shield as float for smooth rendering
